Fix stats calculation for runs with missing start and end events
Summary: Previously, the event log for a run did not contain an event for `PIPELINE_START` or `PIPELINE_SUCCESS | PIPELINE_FAILURE`, then `SqlEventLogStorage.get_stats_for_run` would throw an error.

The error would be thrown because we convert start and end datetime object into floats, but in the case these events were missing, we would attempt to convert `None`.

This diff checks if the times exist, and then pass the datetime to `date_time_as_float`

Test Plan: unit

Reviewers: max

Subscribers: